Add files via upload
In our case and given that we have a stable set of keys, we have obtained a significant improvement in the number of collisions for each function.
a significant improvement of the number of collisions for each hash function proposed by
This work has allowed us to assimilate the notions of the hash function and to understand the importance of the hash function in the context of a stable set of keys.
This work has allowed us to assimilate the notions of hash table and hash functions.
hash functions. One of our main motivations was to avoid having a large number of collisions
a lot of collisions because we understood that it was so perfect or impossible not to have them. We learned a lot of useful techniques like the use of prime numbers, and powers of 2
and powers of 10 as multiplication factors in our hashing algorithms, changing the
changing the size of the ASCII representation of the characters, as well as
the use of bit-level shifting operators to mix the bits well.
Thus we will have a more uniform distribution of the bits, and thus increase the diffusion of the bits in the system.
This way we will have a more uniform distribution of the bits, and thus be able to increase the distribution of the bits in the hash result.